Move common utilities to common import
