+++ src/model_signing/_serialization/serialize_by_file.py
@@ -28,55 +28,6 @@ from model_signing._hashing import file
 from model_signing._serialization import serialization
 
 
-def check_file_or_directory(
-    path: pathlib.Path, *, allow_symlinks: bool = False
-) -> None:
-    """Checks that the given path is either a file or a directory.
-
-    There is no support for sockets, pipes, or any other operating system
-    concept abstracted as a file.
-
-    Furthermore, this would raise if the path is a broken symlink, if it doesn't
-    exists or if there are permission errors.
-
-    Args:
-        path: The path to check.
-        allow_symlinks: Controls whether symbolic links are included. If a
-          symlink is present but the flag is `False` (default) the
-          serialization would raise an error.
-
-    Raises:
-        ValueError: The path is neither a file or a directory, or the path
-          is a symlink and `allow_symlinks` is false.
-    """
-    if not allow_symlinks and path.is_symlink():
-        raise ValueError(
-            f"Cannot use '{path}' because it is a symlink. This"
-            " behavior can be changed with `allow_symlinks`."
-        )
-    if not (path.is_file() or path.is_dir()):
-        raise ValueError(
-            f"Cannot use '{path}' as file or directory. It could be a"
-            " special file, it could be missing, or there might be a"
-            " permission issue."
-        )
-
-
-def should_ignore(
-    path: pathlib.Path, ignore_paths: Iterable[pathlib.Path]
-) -> bool:
-    """Determines if the provided path should be ignored.
-
-    Args:
-        path: The path to check.
-        ignore_paths: The paths to ignore while serializing a model.
-
-    Returns:
-        Whether or not the provided path should be ignored.
-    """
-    return any(path.is_relative_to(ignore_path) for ignore_path in ignore_paths)
-
-
 class FilesSerializer(serialization.Serializer):
     """Generic file serializer.
 
@@ -137,8 +88,12 @@ class FilesSerializer(serialization.Serializer):
         # with `pathlib.Path.walk` for a clearer interface, and some speed
         # improvement.
         for path in itertools.chain((model_path,), model_path.glob("**/*")):
-            check_file_or_directory(path, allow_symlinks=self._allow_symlinks)
-            if path.is_file() and not should_ignore(path, ignore_paths):
+            serialization.check_file_or_directory(
+                path, allow_symlinks=self._allow_symlinks
+            )
+            if path.is_file() and not serialization.should_ignore(
+                path, ignore_paths
+            ):
                 paths.append(path)
 
         manifest_items = []
